Moved get_gdal_version to gdal_utils
